infra/scripts/fabric/requirements.txt
azure-identity>=1.12.0
azure-mgmt-eventhub>=10.1.0
azure-kusto-data
azure-kusto-ingest
pandas
azure-eventhub>=5.11.0
requests>=2.28.0